Start update
Added "remove the cartridge" screen and updated code to display that screen.
Also changed amount of ram reset so that input bytes are not reset.
Moved input code to a macro so that it can be used in ram code.
Created macro for a short delay (just decrements, doesn't wait for anything)
Updated menu_start so that it jumps to the entry point once A is pressed